Manage td-agent.conf file with a fully qualified path
When managing a File resource with Puppet, it's safer to define the
source with a fully qualified path rather than a relative path using
'content'.

In some cases, /etc/puppet/fluentd is actually a symlink to somewhere
else and Puppet is not able to find the actual file if using relative
path.