fix service issues on RHEL 7
RHEL 7 uses systemd for managing services, so setting service_provider to
"redhat" by default results in an error.  This change (a) sets service_provider
to undef in all cases, and (b) exposes a service_provider parameter in init.pp
to allow this to be set if necessary by a calling manifest.

Closes #14.